section6new_self/pyproject.toml
[tool.poetry]
name = "section6new_self"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.2.2"
flask-smorest = "^0.39.0"
flask-sqlalchemy = "^3.0.0"
python-dotenv = "^0.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
